Allneato ping pong eventi per visualizzazione rate anagrafica frontend utente
1. Allneato ping pong eventi per visualizzazione rate anagrafica frontend utente
2. Sistemato errore nel widget budget preventivo in dashboard che non includeva eventuali saldi nelle operazioni di calcolo

+++ app/Http/Controllers/Gestionale/PianiConti/PianoContiController.php
@@ -180,8 +180,8 @@ class PianoContiController extends Controller
             ->orderBy('ragione_sociale')
             ->get(['id', 'ragione_sociale']);
         
-        // AGGIUNGI QUESTA RIGA: Calcolo super-veloce del totale preventivo
-        $totalePreventivo = Conto::where('piano_conto_id', $pianoConto->id)->sum('importo');
+        // Cast a (int) per forzare il tipo numerico ed evitare il warning di Vue
+        $totalePreventivo = (int) Conto::where('piano_conto_id', $pianoConto->id)->sum('importo');
 
         return Inertia::render('gestionale/pianiDeiConti/conti/ContiNew', [
             'condominio' => [
@@ -195,7 +195,7 @@ class PianoContiController extends Controller
             'pianoConti' => new PianoDeiContiResource($pianoConto),
             'conti'      => ContoResource::collection($conti),
             'fornitori'  => $fornitori,
-            'totalePreventivo' => $totalePreventivo,  // ← AGGIUNTO
+            'totalePreventivo' => $totalePreventivo, 
         ]);
     }
